requirements.txt
aiohttp==3.10.10
beautifulsoup4==4.14.2
discord.py==2.4.0
Flask==3.1.2
numpy==2.3.4
opencv_python==4.10.0.84
Pillow==12.0.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
python-dotenv==1.2.1
Requests==2.32.5
scipy==1.16.3
Werkzeug==3.1.3
windows_toasts==1.3.1
